Sprint 3.5 - End

+++ H2Projekt.Application/Commands/Rooms/CreateRoomTypeCommand.cs
@@ -5,6 +5,6 @@
         public string Name { get; set; }
         public string Description { get; set; }
         public int MaxOccupancy { get; set; }
-        public int PricePerNight { get; set; }
+        public decimal PricePerNight { get; set; }
     }
 }
