+++ H2Projekt.Application/Dto/Rooms/RoomTypeDto.cs
@@ -8,7 +8,7 @@ namespace H2Projekt.Application.Dto.Rooms
         public string Name { get; set; } = default!;
         public string? Description { get; set; }
         public int MaxOccupancy { get; set; }
-        public int PricePerNight { get; set; }
+        public decimal PricePerNight { get; set; }
 
         public RoomTypeDto(RoomType roomType)
         {
