+++ H2Projekt.Domain/RoomType.cs
@@ -5,11 +5,11 @@
         public string Name { get; private set; } = default!;
         public string? Description { get; private set; }
         public int MaxOccupancy { get; private set; }
-        public int PricePerNight { get; private set; }
+        public decimal PricePerNight { get; private set; }
 
         public RoomType() { }
 
-        public RoomType(string name, string? description, int maxOccupancy, int pricePerNight)
+        public RoomType(string name, string? description, int maxOccupancy, decimal pricePerNight)
         {
             Name = name;
             Description = description;
@@ -17,7 +17,7 @@
             PricePerNight = pricePerNight;
         }
 
-        public void UpdateDetails(string name, string? description, int maxOccupancy, int pricePerNight)
+        public void UpdateDetails(string name, string? description, int maxOccupancy, decimal pricePerNight)
         {
             Name = name;
             Description = description;
